+++ tests/PhpSpreadsheetTests/Calculation/MathTrigTest.php
@@ -281,10 +281,8 @@ class MathTrigTest extends TestCase
      */
     public function testMINVERSE($expectedResult, ...$args)
     {
-        $this->markTestIncomplete('TODO: This test should be fixed');
-
         $result = MathTrig::MINVERSE(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
+        self::assertEquals($expectedResult, $result, null, 1E-8);
     }
 
     public function providerMINVERSE()
@@ -299,10 +297,8 @@ class MathTrigTest extends TestCase
      */
     public function testMMULT($expectedResult, ...$args)
     {
-        $this->markTestIncomplete('TODO: This test should be fixed');
-
         $result = MathTrig::MMULT(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
+        self::assertEquals($expectedResult, $result, null, 1E-8);
     }
 
     public function providerMMULT()